Errores de mi PR
