+++ git-test/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Home from './components/Home';
-import Login from './components/Login';
+import Home from './components/Home/Home.tsx';
+import Login from './components/Login/Login.tsx';
 
 const App: React.FC = () => {
   return (
